Refactor transform logic to common and restore as a "stage" (#131)
We're using the trick of manifest.Append to create a deep-copy of
r.config, in anticipation of creating the manifests dynamically
according to spec.version. Either way, the transform function can keep
its signature to act as a stage, mutating the passed-by-reference
manifest to a transformed one.

Further, we've simplified the signature of the platform extensions
interface. It may expand once we have a real use case for it.