Add Spec.Version to the API, but not the CRD yet (#135)
* Add Spec.Version to the API, but not the CRD yet

This should be a strict refactoring. Since it's not in the CRD, the
operator will always use the latest embedded version.

The main benefit here is a single, logical place to maintain the
mapping of API type (serving or eventing) to a specific directory
beneath kodata.

* More docs and prevent FinalizeKind from returning an error

* Ignore status.version when calculating target version

If spec.version isn't set, we return latest, regardless of
status.version