Introduce Stages abstraction to simplify manifest post-processing (#138)
* Introduce Stages abstraction to simplify manifest post-processing

This enables all of the stages except the transform functions (which
are the only ones requiring downcasting of the KComponents) to be
moved out of the reconcilers into a common package.

It'll also be handy when obsolete resources are calculated from
diff'ing the target and installed manifests as all of the stages must
be applied to both.

Similarly, the ingress determination should be added to the stages in
FinalizeKind and (the forthcoming) deleteObsoleteResources functions,
along with any other stages that might potentially append to the
constructed manifests.

* Better function names and godoc