Further extension enhancements (#155)
* Further extension enhancements

Introduced a NewExtendedController ctor that takes an
ExtensionGenerator argument which can return an Extension tailored to
a specific platform, e.g. OpenShift, GKE or Minikube.

Both the eventing and serving controllers can be extended with
platform-specific behavior in the reconcile, finalize, and transform
phases.

* Simplify Extension interface

Because error is already a part of the Transformer signature, it
doesn't need to be a part of the Extension.Transformers signature. The
error can be returned when the transform is applied, not when created.

* Removed the context helpers and renamed platform -> extension