Common refactor (#160)
* Refactor common.Install

Both serving and eventing controllers can invoke it directly now.

* Refactor common.CheckDeployments

Both serving and eventing controllers can invoke it directly now, but
not without the context being "injected" with the kube client, which I
couldn't possibly be doing correctly.

* Refactor common.DeleteObsoleteResources

Both serving and eventing controllers can invoke it directly now, but
we really need some e2e tests proving that it works!

* Use manifestival client to check deployments

This removes the need to expect a kube client in the interface and
eliminates a lot of deps related to k8s fake lib.

* Included a unit test for the obsolete resource deletion

* Avoiding optimistic lock exception